+++ src/test/factories/index.ts
@@ -1,6 +1,7 @@
 export * from './createTestBaseField';
 export * from './createTestChangemaker';
 export * from './createTestDataProvider';
+export * from './createTestDefaultPermissionGrant';
 export * from './createTestFile';
 export * from './createTestFunder';
 export * from './createTestInitiative';